vue: refactor
