+++ apps/vue/src/components/challenge/Navbar.vue
